➕ added: initial version of removing users from rooms as admins

+++ PWChat/server/src/Commands/LeaveRoomCommand.cpp
@@ -7,6 +7,11 @@ void LeaveRoomCommand::execute(std::shared_ptr<Session> session, const Packet& p
     LeaveRoomRequest req = p.unpackBody<LeaveRoomRequest>();
     std::shared_ptr<Room> room = server.roomManager().getRoom(req.roomId);
 
+    if (req.userId != session->userId() && !room->checkIfAdmin(session->userId())) {
+        std::cerr << "Error - only admins can remove users from rooms." << std::endl;
+        return;
+    }
+
     bool dbResult = server.db().deleteUserRoom(p.header().senderId, room->id());
 
     if (dbResult) {
